Make functions more modular
Put account related function in `battleground` directory and use main for contract
